% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_description.R
\docType{data}
\name{data1}
\alias{data1}
\title{Brincando com dados em pacote}
\format{
a matrix object with 5 rows and 2 columns
}
\usage{
data1
}
\description{
Exemplo de como colocar dados em pacotes, este dado pode ser qualquer objeto importado para
dentro do pacote
}
\references{
\url{https://www.scielo.br/scielo.php?script=sci_arttext&pid=S1679-62252021000100211}
}
\keyword{datasets}
